requirements.txt
Flask==2.2.1
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2===2.9.6
Werkzeug===2.3.4
PyJWT ===2.7.0